Update README.md
Mainly changes to make it clear that Backdrop projects do not have to be hosted in Backdrop Contrib and that GitHub.com doesn't have the same license requirements as Drupal.org.  Users should be allowed to commit images, fonts, and code using licenses that allow the assets to be distributed within a GPLv2 project.